BUG: When using grid command
bsub command line, used at UNC for killdevil, must contain the '"' character. This was creating issues when writing the python script. We escape the '"' character that is contained in this command line to avoid this problem.
Additionally, when saving and loading this command line, it was split because it contained the '=' sign. We concatenate the string list back to its original form so that it is loaded correctly.