ENH: DTIAtlasBuilder extension superbuild is hidden to Slicer
When DTIAtlasBuilder is built with the Superbuild process, DTIAtlasBuilder installs all the compiled tools in an install directory. This allows to hide the superbuild directory structure to Slicer when creating an extension. This is necessary as DTIAtlasBuilder compiles its own versions of ITK/SlicerExecutionModel. When packaging extensions, the configuration resets the Slicer library paths. In the case of extensions that build their own versions of the libraries, it creates a problem. To avoid that issue, we package directly in the superbuild directory.